backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.6.0
neo4j==5.26.0
redis==5.2.0
httpx==0.28.1
python-multipart==0.0.19
email-validator==2.2.0
